backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.6
groq==0.13.1
python-dotenv==1.0.1
supabase==2.11.0
google-generativeai==0.8.4
